add dynamic boolean expressions * InList * InSubquery
